Release 1-15

+++ backend/apps/secret/routers.py
@@ -21,21 +21,23 @@ __maintainer__ = "Teamlock Project"
 __email__ = "contact@teamlock.io"
 __doc__ = ''
 
+from fastapi import APIRouter, Depends, status, Body, Request, BackgroundTasks
+from toolkits.history import create_history, create_notification
 from toolkits.utils import check_password_complexity
-from fastapi import APIRouter, Depends, status, Body
 from apps.workspace.models import Workspace, Share
 from apps.config.schema import PasswordPolicySchema
 from toolkits.workspace import WorkspaceUtils
+from toolkits.secret import SecretUtils
 from fastapi.exceptions import HTTPException
 from apps.auth.tools import get_current_user
-from toolkits.history import create_history
 from mongoengine.queryset.visitor import Q
 from apps.auth.schema import LoggedUser
 from toolkits.crypto import CryptoUtils
 from fastapi.responses import Response
 from apps.folder.models import Folder
+from apps.workspace.models import Workspace
 from .models import Login, Secret
-from datetime import datetime
+from datetime import date, datetime
 from settings import settings
 from typing import Union
 from . import schema
@@ -97,11 +99,12 @@ async def search_secrets(
 @router.get(
     path="/global/search",
     summary="Search on all workspaces",
-    response_model=list[schema.BankSchema] | list[schema.ServerSchema] | list[schema.LoginSchema] | list[schema.PhoneSchema]
+    response_model=list[schema.BankSchema] | list[schema.ServerSchema] | list[schema.LoginSchema] | list[schema.PhoneSchema] | schema.LoginSchema
 )
 async def global_search_keys(
-    search: str,
-    category: str,
+    search: str = "",
+    category: str = "login",
+    package_name: str = "",
     user: LoggedUser = Depends(get_current_user)
 ):
     workspaces = list(Workspace.objects(owner=user.id))
@@ -110,11 +113,17 @@ async def global_search_keys(
     shares = list(Share.objects(shared_query))
     workspaces.extend([s.workspace for s in shares])
 
-    keys: list = []
+    secrets: list = []
     for workspace in workspaces:
-        keys.extend(WorkspaceUtils.search(workspace.pk, search, user, category))
+        secrets.extend(WorkspaceUtils.search(workspace.pk, search, user, category, package_name=package_name))
     
-    return keys
+    if len(secrets) == 0:
+        return []
+
+    if package_name:
+        return secrets[0]
+
+    return secrets
 
 
 @router.get(
@@ -124,13 +133,16 @@ async def global_search_keys(
 )
 async def get_secret(
     secret_id: str,
+    request: Request,
+    background_task: BackgroundTasks,
     user: LoggedUser = Depends(get_current_user)
 ):
     try:
         secret = Secret.objects(pk=secret_id).get()
         secret_schema = secret.schema()
 
-        workspace, sym_key = WorkspaceUtils.get_workspace(secret.folder.workspace.pk, user)
+        workspace_pk = secret.folder.workspace.pk if secret.folder is not None else secret.trash.workspace.pk
+        workspace, sym_key = WorkspaceUtils.get_workspace(workspace_pk, user)
 
         decrypted_sym_key = CryptoUtils.rsa_decrypt(
             sym_key,
@@ -144,16 +156,30 @@ async def get_secret(
             get_protected_fields=True
         )
 
+        action = f"in folder {secret.folder.name}" if secret.folder is not None else "in trash"
+
         create_history(
             user=user.in_db.email,
             workspace=workspace.name,
             workspace_owner=workspace.owner.email,
-            action=f"Retreive secret for secret {decrypted_secret.name.value} in folder {secret.folder.name}"
+            action=f"Retreive secret for secret {decrypted_secret.name.value} {action}"
         )
 
+        if user.email != decrypted_secret.created_by:
+            create_notification(
+                user=user.id,
+                secret=secret,
+                request=request,
+                mail=True,
+                background_task=background_task
+            )
+
         logger.info(f"[SECRET][{str(workspace.pk)}][{workspace.name}] {user.in_db.email} retreive secret {decrypted_secret.name.value}")
-        
-        # tmp = decrypted_secret.dict()
+        if decrypted_secret.folder is not None:
+            decrypted_secret.folder_name = Folder.objects(pk=decrypted_secret.folder).get().name
+        else:
+            decrypted_secret.folder_name = ""
+        decrypted_secret.workspace_name = workspace.name
         return decrypted_secret
 
     except Secret.DoesNotExist:
@@ -174,12 +200,6 @@ async def create_secret(
 ) -> str:
     try:
         folder: Folder = Folder.objects(pk=schema.secret.folder).get()
-        if folder.in_trash or folder.is_trash:
-            raise HTTPException(
-                status_code=status.HTTP_400_BAD_REQUEST,
-                detail="Can't create a secret inside the trash"
-            )
-
         workspace, sym_key = WorkspaceUtils.get_workspace(folder.workspace.pk, user)
 
         # Check if user is allowed to create a key in this workspace
@@ -196,6 +216,7 @@ async def create_secret(
             
         encrypted_secret = WorkspaceUtils.encrypt_secret(user, sym_key, schema.secret)
         encrypted_secret.folder = folder
+        encrypted_secret.package_name = schema.package_name
 
         encrypted_secret.created_by = user.in_db
         encrypted_secret.updated_by = user.in_db
@@ -230,11 +251,12 @@ async def move_key(
 ):
     try:
         secret: Secret = Secret.objects(pk=secret_id).get()
+        workspace, _ = WorkspaceUtils.get_workspace(secret.folder.workspace.pk if secret.trash is None else secret.trash.workspace.pk, user)
+        WorkspaceUtils.have_rights(workspace, user)
         new_folder: Folder = Folder.objects(pk=folder_id).get()
 
-        workspace, _ = WorkspaceUtils.get_workspace(secret.folder.workspace.pk, user)
-        WorkspaceUtils.have_rights(workspace, user)
-
+        if secret.trash is not None:
+            secret.trash = None
         secret.folder = new_folder
         secret.save()
 
@@ -264,10 +286,10 @@ async def update_secret(
 ) -> None:
     try:
         secret: Secret = Secret.objects(pk=secret_id).get()
-        if secret.folder.in_trash or secret.folder.is_trash:
+        if secret.folder is None:
             raise HTTPException(
-                status_code=status.HTTP_400_BAD_REQUEST,
-                detail="Can't edit a secret inside the trash"
+                status_code=status.HTTP_403_FORBIDDEN,
+                detail="Secret is in trash"
             )
         workspace, sym_key = WorkspaceUtils.get_workspace(secret.folder.workspace.pk, user)
 
@@ -282,10 +304,26 @@ async def update_secret(
         if policy:
             check_password_complexity(policy, schema.secret)
 
+        decrypted_sym_key = CryptoUtils.rsa_decrypt(
+            sym_key,
+            user.in_db.private_key,
+            CryptoUtils.decrypt_password(user)
+        )
+        decrypted_secret = WorkspaceUtils.decrypt_secret(
+            decrypted_sym_key,
+            secret.schema(),
+            get_protected_fields=True
+        )
+
         encrypted_secret: Secret = WorkspaceUtils.encrypt_secret(user, sym_key, schema.secret)
+
+        encrypted_secret.check_changes(decrypted_secret, schema.secret)
 
         encrypted_secret.folder = secret.folder
         encrypted_secret.pk = secret.pk
+
+        if not encrypted_secret.created_by:
+            encrypted_secret.created_by = user.in_db
 
         encrypted_secret.updated_by = user.in_db
         encrypted_secret.save()
@@ -323,7 +361,13 @@ async def delete_secret(
 ) -> None:
     try:
         secret: Secret = Secret.objects(pk=secret_id).get()
-        workspace: Workspace = secret.folder.workspace
+
+        if secret.folder is not None:
+            raise HTTPException(
+                status_code= status.HTTP_400_BAD_REQUEST,
+                detail="You have to first put the secret in the trash to delete it"
+            )
+        workspace: Workspace = secret.trash.workspace
         WorkspaceUtils.have_rights(workspace, user)
         secret.delete()
 
@@ -331,7 +375,7 @@ async def delete_secret(
             user=user.email,
             workspace=workspace.name,
             workspace_owner=workspace.owner.email,
-            action=f"Delete secret {secret.name.value} in folder {secret.folder.name}"
+            action=f"Delete secret {secret.name.value} in trash"
         )
 
         logger.info(f"[SECRET][{str(workspace.pk)}][{workspace.name}] {user.in_db.email} delete secret {secret.name.value}")
@@ -341,3 +385,81 @@ async def delete_secret(
             status_code=status.HTTP_404_NOT_FOUND,
             detail="Secret not found"
         )
+
+
+
+@router.delete(
+    path="/{secret_id}/trash",
+    summary="Move a secret to the trash",
+    status_code=status.HTTP_204_NO_CONTENT
+)
+async def move_to_trash_secret(
+    secret_id: str,
+    user: LoggedUser = Depends(get_current_user)
+) -> None:
+    try:
+        secret: Secret = Secret.objects(pk=secret_id).get()
+        if secret.trash is not None:
+            raise HTTPException(
+                status_code = status.HTTP_400_BAD_REQUEST,
+                detail = "The secret is alreay in the trash"
+            )
+
+        workspace : Workspace = secret.folder.workspace
+        WorkspaceUtils.have_rights(workspace, user)
+        trash : Trash = WorkspaceUtils.get_trash_folder(workspace)
+
+        SecretUtils.move_to_trash(secret, trash)
+
+        create_history(
+            user=user.email,
+            workspace=workspace.name,
+            workspace_owner=workspace.owner.email,
+            action=f"Move secret {secret.name.value} in the trash"
+        )
+
+        logger.info(f"[SECRET][{str(workspace.pk)}][{workspace.name}] {user.in_db.email} move secret {secret.name.value} to trash")
+    
+    except Secret.DoesNotExist:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail="Secret not found"
+        )
+
+@router.patch(
+    path="/{secret_id}/restore",
+    summary = "Restore a secret from the trash",
+    status_code = status.HTTP_204_NO_CONTENT
+)
+async def restore(
+    secret_id: str,
+    user : LoggedUser = Depends(get_current_user)
+) -> None:
+    try:
+        secret: Secret = Secret.objects(pk=secret_id).get()
+    except Secret.DoesNotExist:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail="Secret not found"
+        )
+
+    if secret.trash is None:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail = "The secret is not in the trash"    
+        )
+
+    workspace : Workspace = secret.trash.workspace
+    WorkspaceUtils.have_rights(workspace, user)
+    SecretUtils.restore(secret, workspace,user)
+
+    create_history(
+        user=user.email,
+        workspace=workspace.name,
+        workspace_owner=workspace.owner.email,
+        action=f"Restored secret {secret.name.value}"
+    )
+
+    logger.info(f"[SECRET][{str(workspace.pk)}][{workspace.name}] {user.in_db.email} restored secret {secret.name.value}")
+
+
